:bug: fix but where you cannot sell / buy if you don't choose stock and made it so you can't buy more than your buying power

+++ app/controllers/api/holdings_controller.rb
@@ -25,7 +25,7 @@ class Api::HoldingsController < ApplicationController
 
     def update
         @holding = current_user.holdings.find(params[:id])
-        @holding.quantity += params[:holding][:quantity].to_i
+        @holding.quantity += params[:holding][:quantity].to_f
         if @holding.save
             render :show
         else
